Refactor the semtype class hierachy
