+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/types/semtype/SemType.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com).
+ *  Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
  *
  *  WSO2 LLC. licenses this file to you under the Apache License,
  *  Version 2.0 (the "License"); you may not use this file except
@@ -11,213 +11,40 @@
  *  Unless required by applicable law or agreed to in writing,
  *  software distributed under the License is distributed on an
  *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
+ *  KIND, either express or implied. See the License for the
  *  specific language governing permissions and limitations
  *  under the License.
  */
 
 package io.ballerina.runtime.api.types.semtype;
 
-import io.ballerina.runtime.internal.types.BSemTypeWrapper;
 import io.ballerina.runtime.internal.types.semtype.PureSemType;
 
-import java.lang.ref.Reference;
-import java.lang.ref.WeakReference;
-import java.util.Arrays;
-import java.util.Map;
-import java.util.Objects;
-import java.util.WeakHashMap;
+public interface SemType extends BasicTypeBitSet {
 
-import static io.ballerina.runtime.api.types.semtype.BasicTypeCode.CODE_UNDEF;
-
-/**
- * Runtime representation of SemType.
- *
- * @since 2201.10.0
- */
-public abstract sealed class SemType implements BasicTypeBitSet permits BSemTypeWrapper, PureSemType {
-
-    private static final SubType[] EMPTY_SUBTYPE_DATA = new SubType[0];
-    private static final int CACHEABLE_TYPE_MASK = (~BasicTypeCode.BASIC_TYPE_MASK) & ((1 << (CODE_UNDEF + 1)) - 1);
-
-    private final int all;
-    private final int some;
-    private final SubType[] subTypeData;
-
-    private Integer hashCode;
-
-    private final TypeCheckResultCache resultCache;
-
-    protected SemType(int all, int some, SubType[] subTypeData) {
-        this.all = all;
-        this.some = some;
-        this.subTypeData = subTypeData;
-        if ((some & CACHEABLE_TYPE_MASK) != 0) {
-            this.resultCache = new TypeCheckResultCache();
-        } else {
-            this.resultCache = TypeCheckResultCache.EMPTY;
-        }
-    }
-
-    protected SemType(int all) {
-        this(all, 0, EMPTY_SUBTYPE_DATA);
-    }
-
-    protected SemType(SemType semType) {
-        this(semType.all(), semType.some(), semType.subTypeData());
-    }
-
-    public static SemType from(int all, int some, SubType[] subTypeData) {
+    static SemType from(int all, int some, SubType[] subTypeData) {
         return new PureSemType(all, some, subTypeData);
     }
 
-    public static SemType from(int all) {
+    static SemType from(int all) {
         return new PureSemType(all);
     }
 
-    @Override
-    public String toString() {
-        return SemTypeHelper.stringRepr(this);
-    }
+    int all();
 
-    public final int all() {
-        return all;
-    }
+    int some();
 
-    public final int some() {
-        return some;
-    }
+    SubType[] subTypeData();
 
-    public final SubType[] subTypeData() {
-        return subTypeData;
-    }
+    CachedResult cachedSubTypeRelation(SemType other);
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof SemType semType)) {
-            return false;
-        }
-        return all() == semType.all() && some() == semType.some() &&
-                Objects.deepEquals(subTypeData, semType.subTypeData);
-    }
+    void cacheSubTypeRelation(SemType other, boolean result);
 
-    @Override
-    public int hashCode() {
-        Integer result = hashCode;
-        if (result == null) {
-            synchronized (this) {
-                result = hashCode;
-                if (result == null) {
-                    hashCode = result = computeHashCode();
-                }
-            }
-        }
-        return result;
-    }
+    SubType subTypeByCode(int code);
 
-    private int computeHashCode() {
-        return Objects.hash(all(), some(), Arrays.hashCode(subTypeData));
-    }
-
-    private boolean shouldCache() {
-        return (some() & CACHEABLE_TYPE_MASK) != 0;
-    }
-
-    enum CachedResult {
+    public enum CachedResult {
         TRUE,
         FALSE,
         NOT_FOUND
-    }
-
-    CachedResult cachedSubTypeRelation(SemType other) {
-        if (!shouldCache()) {
-            return CachedResult.NOT_FOUND;
-        }
-        if (this == other) {
-            return CachedResult.TRUE;
-        }
-        return resultCache.getCachedResult(other);
-    }
-
-    void cacheSubTypeRelation(SemType other, boolean result) {
-        if (shouldCache()) {
-            resultCache.cacheResult(other, result);
-            assert isValidCacheState(other, result) : "Invalid cache state";
-        }
-    }
-
-    private boolean isValidCacheState(SemType other, boolean result) {
-        CachedResult cachedResult = cachedSubTypeRelation(other);
-        return cachedResult == CachedResult.NOT_FOUND ||
-                cachedResult == (result ? CachedResult.TRUE : CachedResult.FALSE);
-    }
-
-    public final SubType subTypeByCode(int code) {
-        if ((some() & (1 << code)) == 0) {
-            return null;
-        }
-        int someMask = (1 << code) - 1;
-        int some = some() & someMask;
-        return subTypeData()[Integer.bitCount(some)];
-    }
-
-    private static sealed class TypeCheckResultCache {
-
-        private static final TypeCheckResultCache EMPTY = new EmptyTypeCheckResultCache();
-        // make this an int
-        private final Map<TypeCheckCacheKey, Boolean> cache = new WeakHashMap<>();
-
-        public void cacheResult(SemType semType, boolean result) {
-            cache.put(TypeCheckCacheKey.from(semType), result);
-        }
-
-        public CachedResult getCachedResult(SemType semType) {
-            Boolean cachedData = cache.get(TypeCheckCacheKey.from(semType));
-            if (cachedData == null) {
-                return CachedResult.NOT_FOUND;
-            }
-            return cachedData ? CachedResult.TRUE : CachedResult.FALSE;
-        }
-    }
-
-    private static final class EmptyTypeCheckResultCache extends TypeCheckResultCache {
-
-        @Override
-        public void cacheResult(SemType semType, boolean result) {
-            throw new UnsupportedOperationException("Empty cache");
-        }
-
-        @Override
-        public CachedResult getCachedResult(SemType semType) {
-            throw new UnsupportedOperationException("Empty cache");
-        }
-    }
-
-    private record TypeCheckCacheKey(Reference<SemType> semtype) {
-
-        static TypeCheckCacheKey from(SemType semType) {
-            return new TypeCheckCacheKey(new WeakReference<>(semType));
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (!(obj instanceof TypeCheckCacheKey other)) {
-                return false;
-            }
-            SemType thisSemType = semtype.get();
-            SemType otherSemType = other.semtype.get();
-            if (thisSemType == null || otherSemType == null) {
-                return false;
-            }
-            return thisSemType == otherSemType;
-        }
-
-        @Override
-        public int hashCode() {
-            return System.identityHashCode(semtype.get());
-        }
     }
 }
